DicomServiceApi removed, all dicom web client methods unified in DICOMWebService
