+++ shanoir-ng-datasets/src/main/java/org/shanoir/ng/dicom/web/service/DICOMWebService.java
@@ -1,14 +1,24 @@
 package org.shanoir.ng.dicom.web.service;
 
+import java.io.File;
+import java.io.IOException;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
 import javax.annotation.PostConstruct;
 
 import org.apache.http.HttpEntity;
+import org.apache.http.client.ClientProtocolException;
 import org.apache.http.client.methods.CloseableHttpResponse;
+import org.apache.http.client.methods.HttpDelete;
 import org.apache.http.client.methods.HttpGet;
 import org.apache.http.client.methods.HttpPost;
+import org.apache.http.entity.ContentType;
+import org.apache.http.entity.mime.MultipartEntityBuilder;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClients;
 import org.apache.http.util.EntityUtils;
+import org.shanoir.ng.shared.exception.ShanoirException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
@@ -17,12 +27,23 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
-import org.springframework.web.multipart.MultipartFile;
 
 @Component
 public class DICOMWebService {
 
 	private static final Logger LOG = LoggerFactory.getLogger(DICOMWebService.class);
+
+	private static final String CONTENT_TYPE_MULTIPART = "multipart/related";
+	
+	private static final String CONTENT_TYPE_DICOM = "application/dicom";
+
+	private static final String CONTENT_TYPE_JSON = "application/json";
+
+	private static final String BOUNDARY = "--import_dicom_shanoir--";
+
+	private static final String CONTENT_TYPE = "&contentType";
+
+	private static final String REJECT_SUFFIX = "/reject/113001%5EDCM";
 
 	private CloseableHttpClient httpClient;
 
@@ -114,19 +135,136 @@ public class DICOMWebService {
 		return null;
 	}
 
-	public String stow(MultipartFile file) {
-		try {
-			HttpPost httpPost = new HttpPost(this.serverURL.substring(0, this.serverURL.indexOf("/studies")));
-			//httpPost.setEntity(null)
-			CloseableHttpResponse response = httpClient.execute(httpPost);
-			HttpEntity entity = response.getEntity();
-			if (entity != null) {
-				return EntityUtils.toString(entity);
-			}
-		} catch (Exception e) {
-			LOG.error(e.getMessage(), e);
+	public void sendDicomFilesToPacs(File directoryWithDicomFiles) throws Exception {
+		if (directoryWithDicomFiles == null || !directoryWithDicomFiles.exists() || !directoryWithDicomFiles.isDirectory()) {
+			throw new ShanoirException("sendDicomFilesToPacs called with null, or file: not existing or not a directory.");
 		}
-		return null;
+		File[] dicomFiles = directoryWithDicomFiles.listFiles();
+		LOG.info("Start: STOW-RS sending " + dicomFiles.length + " dicom files to PACS from folder: " + directoryWithDicomFiles.getAbsolutePath());
+
+		try(CloseableHttpClient httpClient = HttpClients.createDefault()) {
+			MultipartEntityBuilder multipartEntityBuilder = MultipartEntityBuilder.create().setBoundary(BOUNDARY);
+			for (File dicomFile : dicomFiles) {
+				multipartEntityBuilder.addBinaryBody("dcm_upload", dicomFile, ContentType.create(CONTENT_TYPE_DICOM), "filename");
+			}
+			HttpEntity entity = multipartEntityBuilder.build();
+			HttpPost httpPost = new HttpPost(dcm4cheeProtocol + dcm4cheeHost + ":" + dcm4cheePort + dicomWebRS);
+			httpPost.setHeader(HttpHeaders.CONTENT_TYPE, CONTENT_TYPE_MULTIPART+";type="+CONTENT_TYPE_DICOM+";boundary="+BOUNDARY);
+			httpPost.setEntity(entity);
+			CloseableHttpResponse response = httpClient.execute(httpPost);
+			response.getEntity();
+			response.close();
+		} catch (ClientProtocolException e) {
+			LOG.error("ClientProtocolException during upload into pacs",e);
+			throw e;
+		} catch (IOException e) {
+			LOG.error("IOException during upload into pacs",e);
+			throw e;
+		}
+		LOG.info("Finished: STOW-RS sending " + dicomFiles.length + " dicom files to PACS from folder: " + directoryWithDicomFiles.getAbsolutePath());
 	}
 
+	public void deleteDicomFilesFromPacs(String url) throws ShanoirException {
+		String instanceId;
+		String studyId;
+		String serieId;
+		String rejectURL;
+		String deleteUrl;
+
+		if (url.contains("requestType=WADO")) {
+			instanceId = this.extractInstanceUID(url, null);
+			studyId = this.extractStudyUID(url, null);
+			serieId = this.extractSeriesUIDUID(url, null);
+			// http://localhost:8081/dcm4chee-arc/aets/DCM4CHEE/rs/studies//series//instances//reject/113001%5EDCM
+			rejectURL = url.substring(0, url.indexOf("wado?")) + "rs/studies/" + studyId + "/series/" + serieId
+					+ "/instances/" + instanceId + REJECT_SUFFIX;
+			deleteUrl = url.substring(0, url.indexOf("/aets/")) + REJECT_SUFFIX;
+		} else {
+			// /studies/{study}/series/{series}/instances/{instance}/rendered
+			Pattern p  = Pattern.compile(".*//studies//(.*)//series//(.*)//instances//(.*)");
+			Matcher m = p.matcher(url);
+			if (m.find()) {
+				studyId = m.group(1);
+				serieId = m.group(2);
+				instanceId = m.group(3);
+			}
+			rejectURL = url + REJECT_SUFFIX;
+			deleteUrl = url.substring(0, url.indexOf("/aets/")) + REJECT_SUFFIX;
+		}
+
+		try(CloseableHttpClient httpClient = HttpClients.createDefault()) {
+			// STEP 1: Reject from the PACS
+			HttpPost post = new HttpPost(rejectURL);
+			post.setHeader(HttpHeaders.CONTENT_TYPE, CONTENT_TYPE_JSON);
+			CloseableHttpResponse response = httpClient.execute(post);
+			if (response.getStatusLine().getStatusCode() == HttpStatus.NO_CONTENT.value()) {
+				LOG.info("Rejected from PACS: " + url);
+			} else {
+				LOG.error(response.getStatusLine().getStatusCode() + ": Could not reject instance from PACS: "
+						+ response.getStatusLine().getReasonPhrase());
+				throw new ShanoirException("Could not reject instance from PACS: " + rejectURL);
+			}
+			// STEP 2: Delete from the PACS
+			HttpDelete delete = new HttpDelete(deleteUrl);
+			delete.setHeader(HttpHeaders.CONTENT_TYPE, CONTENT_TYPE_JSON);
+			response = httpClient.execute(delete);
+			if (response.getStatusLine().getStatusCode() == HttpStatus.OK.value()) {
+				LOG.info("Deleted from PACS: " + url);
+			} else {
+				LOG.error(response.getStatusLine().getStatusCode() + ": Could not delete instance from PACS: "
+						+ response.getStatusLine().getReasonPhrase());
+				throw new ShanoirException("Could not delete instance from PACS: " + deleteUrl);
+			}
+		} catch (ClientProtocolException e) {
+			throw new ShanoirException("ClientProtocolException during delete from pacs: " + url, e);
+		} catch (IOException e) {
+			throw new ShanoirException("IOException during delete from pacs: " + url, e);
+		}
+	}
+
+	/**
+	 * The instanceUID (== objectUID) is inside the URL string
+	 * and has to be extracted to be used.
+	 * 
+	 * @param url
+	 * @param instanceUID
+	 * @return
+	 */
+	private String extractInstanceUID(String url, String instanceUID) {
+		return extractUidPattern(url, "objectUID", url.indexOf(CONTENT_TYPE) != -1 ? CONTENT_TYPE : null, instanceUID);
+	}
+
+	/**
+	 * The studyID is inside the URL string
+	 * and has to be extracted to be used.
+	 * 
+	 * @param url
+	 * @param defaultUID
+	 * @return the studyUID
+	 */
+	private String extractStudyUID(String url, String studyUID) {
+		return extractUidPattern(url, "studyUID", "&seriesUID", studyUID);
+	}
+
+	/**
+	 * The series UID is inside the URL string
+	 * and has to be extracted to be used.
+	 * 
+	 * @param url
+	 * @param defaultUID
+	 * @return the seriesUID
+	 */
+	private String extractSeriesUIDUID(String url, String seriesUID) {
+		return extractUidPattern(url, "seriesUID", "&objectUID", seriesUID);
+	}
+
+	private String extractUidPattern(String url, String uidName, String endPattern, String defaultValue) {
+		Pattern p  = Pattern.compile(".*" + uidName + "=(.*)" + (endPattern != null ? endPattern + ".*" : ""));
+		Matcher m = p.matcher(url);
+		if (m.find()) {
+			defaultValue = m.group(1);
+		}
+		return defaultValue;
+	}
+	
 }
